# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 企业内部练习与考试系统 API
  version: 1.0.0
  description: |
    MVP 完整接口规格。全局约定见 [00-API设计说明.md](./00-API设计说明.md)。
    分域行为见 01~07 Markdown 文档。
  contact:
    name: Exam System Team
servers:
  - url: /api/v1
    description: 单租户企业内网 API
...
security:
  - sessionAuth: []
paths:
...
  /question-banks/{id}:
    patch:
      tags: [QuestionBank]
      operationId: updateQuestionBank
      summary: 更新题库与开放策略
      x-fr-ids: [FR-QST-01]
      x-page-ids: [AD-05]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /question-banks/{bankId}/categories:
    get:
      tags: [QuestionBank]
      operationId: listCategories
      summary: 分类列表
      x-fr-ids: [FR-QST-02]
      x-page-ids: [AD-06]
      parameters:
        - name: bankId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    post:
      tags: [QuestionBank]
      operationId: createCategory
      summary: 创建分类
      x-fr-ids: [FR-QST-02]
      parameters:
        - name: bankId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          $ref: '#/components/responses/SuccessEmpty'
  /categories/{id}:
    patch:
      tags: [QuestionBank]
      operationId: updateCategory
      summary: 更新分类
      x-fr-ids: [FR-QST-02]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /categories/{categoryId}/knowledge-points:
    get:
      tags: [QuestionBank]
      operationId: listKnowledgePoints
      summary: 知识点列表
      x-fr-ids: [FR-QST-02]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    post:
      tags: [QuestionBank]
      operationId: createKnowledgePoint
      summary: 创建知识点
      x-fr-ids: [FR-QST-02]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          $ref: '#/components/responses/SuccessEmpty'
  /knowledge-points/{id}:
    patch:
      tags: [QuestionBank]
      operationId: updateKnowledgePoint
      summary: 更新知识点
      x-fr-ids: [FR-QST-02]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /question-banks/{bankId}/questions:
    get:
      tags: [QuestionBank]
      operationId: listQuestions
      summary: 题目列表
      x-fr-ids: [FR-QST-03]
      x-page-ids: [AD-06]
      parameters:
        - name: bankId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    post:
      tags: [QuestionBank]
      operationId: createQuestion
      summary: 创建题目
      x-fr-ids: [FR-QST-03, FR-QST-04]
      x-page-ids: [AD-07]
      parameters:
        - name: bankId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          $ref: '#/components/responses/SuccessEmpty'
  /questions/{id}:
    get:
      tags: [QuestionBank]
      operationId: getQuestion
      summary: 题目详情
      x-fr-ids: [FR-QST-03]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    patch:
      tags: [QuestionBank]
      operationId: updateQuestion
      summary: 更新题目元数据（不含内容版本）
      x-fr-ids: [FR-QST-03]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /questions/{id}/versions:
    get:
      tags: [QuestionBank]
      operationId: listQuestionVersions
      summary: 题目版本历史
      x-fr-ids: [FR-QST-03]
      x-page-ids: [AD-07]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
    post:
      tags: [QuestionBank]
      operationId: createQuestionVersion
      summary: 创建不可变新版本
      x-fr-ids: [FR-QST-03, FR-QST-04]
      x-page-ids: [AD-07]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionVersionInput'
      responses:
        '201':
          description: 新版本
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/QuestionVersionAdmin'
  /question-versions/{versionId}:
    get:
      tags: [QuestionBank]
      operationId: getQuestionVersion
      summary: 题目版本详情
      x-fr-ids: [FR-QST-03]
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
...
  /admin/exams/{id}/scores/summary:
    get:
      tags: [Report]
      operationId: getExamScoreSummary
      summary: 双维度统计摘要
      x-fr-ids: [FR-REP-01]
      x-page-ids: [AD-14]
      x-sys-ids: [SYS-06]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/scores/employees:
    get:
      tags: [Report]
      operationId: listExamEmployeeScores
      summary: 员工官方成绩表
      x-fr-ids: [FR-REP-02, FR-SCR-02]
      x-page-ids: [AD-14]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/attempts:
    get:
      tags: [Report]
      operationId: listAdminExamAttempts
      summary: 全部尝试列表
      x-fr-ids: [FR-REP-02]
      x-page-ids: [AD-15]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/attempts/{attemptId}:
    get:
      tags: [Report]
      operationId: getAdminExamAttempt
      summary: 尝试详情（管理端完整）
      x-fr-ids: [FR-REP-02, FR-SCR-03]
      x-page-ids: [AD-15]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/attempts/{attemptId}/void:
    post:
      tags: [Score]
      operationId: voidExamAttempt
      summary: 作废尝试并重算
      x-fr-ids: [FR-SCR-03]
      x-page-ids: [AD-15]
      x-sys-ids: [SYS-04]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeVisibleReason, internalReason]
              properties:
                employeeVisibleReason:
                  type: string
                internalReason:
                  type: string
                refundAttempt:
                  type: boolean
      responses:
        '200':
          $ref: '#/components/responses/SuccessEmpty'
  /admin/exams/{id}/exports:
    post:
      tags: [Report]
      operationId: createExamExport
      summary: 创建异步导出
      x-fr-ids: [FR-REP-03]
      x-page-ids: [AD-14]
      x-sys-ids: [SYS-06]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [xlsx]
                  default: xlsx
                includeSheets:
                  type: array
                  items:
                    type: string
                    enum: [officialSummary, allAttempts]
      responses:
        '202':
          description: 导出任务已创建
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          jobId:
                            type: string
                          status:
                            $ref: '#/components/schemas/AsyncJobStatus'
  /admin/exports/{jobId}:
    get:
      tags: [Report]
      operationId: getExportJob
      summary: 查询导出任务
      x-fr-ids: [FR-REP-03]
      x-page-ids: [AD-14]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 导出任务状态
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExportJob'
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 亦可用 HttpOnly Session Cookie，二者等价
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    SuccessEmpty:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ResponseMeta:
      type: object
      required: [serverNow, requestId]
      properties:
        serverNow:
          type: string
          format: date-time
        requestId:
          type: string
...
    QuestionType:
      type: string
      enum: [singleChoice, multipleChoice, trueFalse]
    QuestionVersionInput:
      type: object
      required: [type, stem, options, standardAnswer]
      properties:
        type:
          $ref: '#/components/schemas/QuestionType'
        stem:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              text:
                type: string
        standardAnswer:
          type: array
          items:
            type: string
        explanation:
          type: string
        categoryId:
          type: string
        knowledgePointId:
          type: [string, 'null']
        difficulty:
          type: string
          enum: [easy, medium, hard]
        defaultScore:
          type: number
          format: double
    QuestionVersionAdmin:
      allOf:
        - $ref: '#/components/schemas/QuestionVersionInput'
        - type: object
          properties:
            id:
              type: string
            questionId:
              type: string
            versionNo:
              type: integer
            status:
              type: string
              enum: [active, disabled]
            createdAt:
              type: string
              format: date-time
...
    AsyncJobStatus:
      type: string
      enum: [pending, processing, completed, failed, cancelled, expired]
    ExportJob:
      type: object
      properties:
        jobId:
          type: string
        status:
          $ref: '#/components/schemas/AsyncJobStatus'
        downloadUrl:
          type: [string, 'null']
        expiresAt:
          type: [string, 'null']
          format: date-time